Add React Router and individual blog post pages
